Chicca della Apple

+++ app/src/main/java/com/example/cogor/navigationdrawer/CartListAdapter.java
@@ -23,12 +23,12 @@ import java.util.ArrayList;
  * Created by cogor on 19/09/2017.
  */
 
-public class CartListAdapter extends BaseAdapter{
+public class CartListAdapter extends BaseAdapter {
 
     ArrayList<CartItem> cartItems;
     Activity activity;
-    public CartListAdapter(ArrayList<CartItem> cartItems, Activity activity)
-    {
+
+    public CartListAdapter(ArrayList<CartItem> cartItems, Activity activity) {
         this.cartItems = cartItems;
         this.activity = activity;
     }
@@ -50,7 +50,7 @@ public class CartListAdapter extends BaseAdapter{
 
     @Override
     public View getView(int position, View convertView, ViewGroup parent) {
-            View view = activity.getLayoutInflater().inflate(R.layout.cart_list_item, null);
+        View view = activity.getLayoutInflater().inflate(R.layout.cart_list_item, null);
         final CartItem item = cartItems.get(position);
 
         TextView itemName = (TextView) view.findViewById(R.id.itemCartName);
@@ -60,7 +60,7 @@ public class CartListAdapter extends BaseAdapter{
         Button decrement = (Button) view.findViewById(R.id.decrementCart);
         itemName.setText(item.getProdName());
         itemAmount.append(item.getAmount());
-        itemQuantity.append(Integer.toString(item.getQuantity()));
+        itemQuantity.append(Double.toString(item.getQuantity()));
         increment.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -71,16 +71,16 @@ public class CartListAdapter extends BaseAdapter{
                 SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(activity.getApplicationContext());
                 String username = prefs.getString("Username", null);
 
-                String[] toCheck = new String[] {username, item.getId()};
+                String[] toCheck = new String[]{username, item.getId()};
 
                 String[] toUpdate = new String[]{"quantity", "singleamount"};
-                Cursor infoCursor = db.query(DbCart.CartInit.TABLE_NAME , toUpdate, "username = ? AND prodid = ?" , toCheck, null, null, null);
+                Cursor infoCursor = db.query(DbCart.CartInit.TABLE_NAME, toUpdate, "username = ? AND prodid = ?", toCheck, null, null, null);
                 infoCursor.moveToFirst();
 
                 int currentQuantity = infoCursor.getInt(infoCursor.getColumnIndex(DbCart.CartInit.COLUMN_NAME_QUANTITY));
                 double currentAmount = infoCursor.getDouble(infoCursor.getColumnIndex(DbCart.CartInit.COLUMN_NAME_SINGLEAMOUNT));
-                contentValues.put(DbCart.CartInit.COLUMN_NAME_QUANTITY, currentQuantity +1);
-                contentValues.put(DbCart.CartInit.COLUMN_NAME_SINGLEAMOUNT, currentAmount + currentAmount/item.getQuantity());
+                contentValues.put(DbCart.CartInit.COLUMN_NAME_QUANTITY, currentQuantity + 1);
+                contentValues.put(DbCart.CartInit.COLUMN_NAME_SINGLEAMOUNT, currentAmount + currentAmount / item.getQuantity());
                 db.update(DbCart.CartInit.TABLE_NAME, contentValues, "username = ? AND prodid = ?", toCheck);
                 reloadFrag(activity);
             }
@@ -95,10 +95,10 @@ public class CartListAdapter extends BaseAdapter{
                 SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(activity.getApplicationContext());
                 String username = prefs.getString("Username", null);
 
-                String[] toCheck = new String[] {username, item.getId()};
+                String[] toCheck = new String[]{username, item.getId()};
 
                 String[] toUpdate = new String[]{"quantity", "singleamount"};
-                Cursor infoCursor = db.query(DbCart.CartInit.TABLE_NAME , toUpdate, "username = ? AND prodid = ?" , toCheck, null, null, null);
+                Cursor infoCursor = db.query(DbCart.CartInit.TABLE_NAME, toUpdate, "username = ? AND prodid = ?", toCheck, null, null, null);
                 infoCursor.moveToFirst();
 
                 int currentQuantity = infoCursor.getInt(infoCursor.getColumnIndex(DbCart.CartInit.COLUMN_NAME_QUANTITY));
@@ -110,7 +110,7 @@ public class CartListAdapter extends BaseAdapter{
                     return;
                 }
                 contentValues.put(DbCart.CartInit.COLUMN_NAME_QUANTITY, currentQuantity - 1);
-                contentValues.put(DbCart.CartInit.COLUMN_NAME_SINGLEAMOUNT, currentAmount - currentAmount/item.getQuantity());
+                contentValues.put(DbCart.CartInit.COLUMN_NAME_SINGLEAMOUNT, currentAmount - currentAmount / item.getQuantity());
                 db.update(DbCart.CartInit.TABLE_NAME, contentValues, "username = ? AND prodid = ?", toCheck);
                 reloadFrag(activity);
             }
@@ -118,16 +118,16 @@ public class CartListAdapter extends BaseAdapter{
         return view;
     }
 
-    public void reloadFrag(Activity activity)
-    {
+    public void reloadFrag(Activity activity) {
         Fragment currentFragment = activity.getFragmentManager().findFragmentById(R.id.content_frame);
         if (currentFragment instanceof CartFragment) {
-        android.app.FragmentTransaction fragTransaction =   activity.getFragmentManager().beginTransaction();
-        fragTransaction.detach(currentFragment);
-        fragTransaction.attach(currentFragment);
-        fragTransaction.commit();}
+            android.app.FragmentTransaction fragTransaction = activity.getFragmentManager().beginTransaction();
+            fragTransaction.detach(currentFragment);
+            fragTransaction.attach(currentFragment);
+            fragTransaction.commit();
+        }
     }
 
-    }
+}
